Websocket

+++ ProductosMongoRestNet/Controllers/Books/BooksController.cs
@@ -2,6 +2,7 @@
 using ProductosMongoRestNet.Models;
 using ProductosMongoRestNet.Services;
 using ProductosMongoRestNet.Services.Storage;
+using ProductosMongoRestNet.Websocket;
 
 namespace ProductosMongoRestNet.Controllers;
 
@@ -13,11 +14,14 @@ public class BooksController : ControllerBase
     private const string _route = "api/storage";
     private readonly IBooksService _booksService;
     private readonly IFileStorageService _storageService;
+    private readonly WebSocketHandler _webSocketHandler; // Añadimos el WebSocketHandler
 
-    public BooksController(IBooksService booksService, IFileStorageService storageService)
+    public BooksController(IBooksService booksService, IFileStorageService storageService,
+        WebSocketHandler webSocketHandler)
     {
         _booksService = booksService;
         _storageService = storageService;
+        _webSocketHandler = webSocketHandler; // Inyectamos el WebSocketHandler
     }
 
     [HttpGet]
@@ -41,6 +45,11 @@ public class BooksController : ControllerBase
     public async Task<ActionResult<Book>> Create(Book book)
     {
         var savedBook = await _booksService.CreateAsync(book);
+
+        // Enviamos la notificación a todos los clientes conectados
+        await _webSocketHandler.NotifyAllAsync("New book created with id: " + savedBook.Id);
+
+        // Devolvemos la respuesta con el libro creado
         return CreatedAtAction(nameof(GetById), new { id = book.Id }, savedBook);
     }
 
@@ -53,6 +62,9 @@ public class BooksController : ControllerBase
 
         if (updatedBook is null) return NotFound("Book not found with the provided id: " + id);
 
+        // Enviamos la notificación a todos los clientes conectados
+        await _webSocketHandler.NotifyAllAsync("Book updated with id: " + updatedBook.Id);
+
         return Ok(updatedBook);
     }
 
@@ -62,41 +74,43 @@ public class BooksController : ControllerBase
         var deletedBook = await _booksService.DeleteAsync(id);
 
         if (deletedBook is null) return NotFound("Book not found with the provided id: " + id);
-        
+
         // Eliminamos la imagen
         try
         {
-            if (!string.IsNullOrEmpty(deletedBook.Image))
-            {
-                await _storageService.DeleteFileAsync(deletedBook.Image);
-            }
-            return NoContent();
-        } catch (Exception ex)
+            if (!string.IsNullOrEmpty(deletedBook.Image)) await _storageService.DeleteFileAsync(deletedBook.Image);
+        }
+        catch (Exception ex)
         {
             return BadRequest(ex.Message);
         }
+
+        // Eliminamos la notificación a todos los clientes conectados
+        await _webSocketHandler.NotifyAllAsync("Book deleted with id: " + deletedBook.Id);
+
+        // Devolvemos la respuesta con el libro eliminado
+        return NoContent();
     }
-    
+
     [HttpPatch("{id:length(24)}")]
     public async Task<ActionResult> UpdateImage(
         string id,
         [FromForm] IFormFile file)
     {
-        
         // Comprobamos que el fichero no sea nulo
         if (file == null || file.Length == 0)
             return BadRequest("Not file in the request");
-        
+
         // Obtenemos el libro
         var book = await _booksService.GetByIdAsync(id);
-        
+
         // Si el libro no existe, devolvemos un error
         if (book is null) return NotFound("Book not found with the provided id: " + id);
         try
         {
             // Guardamos el fichero
             var fileName = await _storageService.SaveFileAsync(file);
-            
+
             // Actualizamos la URL de la imagen
             // Aquí es cuando debemos decidir si la queremos el no,bre del fichero o la URL
             // Mira el controlador de Storage para ver cómo se hace, yo lo he hecho con el nombre del fichero
@@ -106,8 +120,12 @@ public class BooksController : ControllerBase
             var fileUrl = $"{baseUrl}/{_route}/{fileName}";
             book.Image = fileUrl;
             //book.Image = fileName;
-            return Ok(await _booksService.UpdateAsync(id, book));
 
+            // Avismamos a los clientes conectados
+            await _webSocketHandler.NotifyAllAsync("Book image updated with id: " + book.Id);
+
+            // Devolvemos el libro actualizado con la nueva URL de la imagen
+            return Ok(await _booksService.UpdateAsync(id, book));
         }
         catch (Exception ex)
         {
